[VGNWB-366] citesphere display files
